requirements.txt
qdrant-client 
transformers 
langchain_openai
langchain-groq
ipykernel
sentence-transformers
rank-bm25 
pypdf
PyPDF2
google-genai
streamlit